backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn
